requirement.txt
openai
supabase
pillow
pandas
langfuse
aiohttp
azure-cognitiveservices-speech
python-dotenv
google-genai